Delete screen added to the manager privileges

+++ controller/manager.go
@@ -37,6 +37,26 @@ func ManagerLogin(c *gin.Context) {
 	c.JSON(http.StatusAccepted, gin.H{"status": "true", "error": token})
 }
 
+func GetScreenList(c *gin.Context) {
+	managerUsername := c.GetString("username")
+	var manager models.Manager
+	result := config.DB.Where("username = ?", managerUsername).First(&manager)
+	if result.Error != nil {
+		if result.Error != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"status": "false", "error": result.Error.Error()})
+			return
+		}
+	}
+	var screens []models.Screen
+
+	result = config.DB.Where("cinemas_id = ?", manager.CinemasId).Find(&screens)
+	if result.Error != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"status": "false", "error": result.Error.Error()})
+		return
+	}
+	c.JSON(http.StatusAccepted, gin.H{"status": "true", "screens": screens})
+}
+
 func AddScreen(c *gin.Context) {
 	var screen models.Screen
 	if err := c.ShouldBindJSON(&screen); err != nil {
@@ -45,7 +65,7 @@ func AddScreen(c *gin.Context) {
 	}
 
 	//checking the screen name already exists
-	search := config.DB.First(&models.Screen{Name: screen.Name, CinemasId: screen.CinemasId})
+	search := config.DB.Where("name = ? AND cinemas_id = ?", screen.Name, screen.CinemasId).First(&models.Screen{})
 	if search.RowsAffected != 0 {
 		c.JSON(http.StatusUnauthorized, gin.H{"status": "false", "message": "Screen Name already exists for the cinemas"})
 		return
@@ -69,7 +89,7 @@ func EditScreen(c *gin.Context) {
 	}
 
 	//checking the screen name already exists
-	search := config.DB.Not("id = ?", id).First(&models.Screen{Name: screen.Name, CinemasId: screen.CinemasId})
+	search := config.DB.Not("id = ?", id).Where("name = ? AND cinemas_id = ?", screen.Name, screen.CinemasId).First(&models.Screen{})
 	if search.RowsAffected != 0 {
 		c.JSON(http.StatusUnauthorized, gin.H{"status": "false", "message": "Screen Name already exists for the cinemas"})
 		return
@@ -85,6 +105,27 @@ func EditScreen(c *gin.Context) {
 
 func DeleteScreen(c *gin.Context) {
 	id := c.Param("id")
+	var manager models.Manager
+
+	//getting manager details
+	managerUsername := c.GetString("username")
+	getManager := config.DB.Where("username = ?", managerUsername).First(&manager)
+	if getManager.RowsAffected == 0 {
+		c.JSON(http.StatusUnauthorized, gin.H{"status": "false", "message": "Unautherised access"})
+		return
+	}
+	var screen models.Screen
+	//verifiying the manager is deleting the screen on his theathre
+	verify := config.DB.Where("id = ?", id).First(&screen)
+	if verify.RowsAffected == 0 {
+		c.JSON(http.StatusUnauthorized, gin.H{"status": "false", "message": "Already deleted"})
+		return
+	}
+
+	if manager.CinemasId != screen.CinemasId {
+		c.JSON(http.StatusUnauthorized, gin.H{"status": "false", "message": "Unautherised access"})
+		return
+	}
 	result := config.DB.Delete(&models.Screen{}, id)
 	if result.Error != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"status": "false", "error": result.Error.Error()})
